requirements.txt
absl-py==2.0.0
accelerate==0.24.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.4.2
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.7
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.2.0
contourpy==1.2.0
ctransformers==0.2.27
cycler==0.12.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
diskcache==5.6.3
distro==1.8.0
docker-pycreds==0.4.0
executing==2.0.1
exllama==0.1.0
fastjsonschema==2.19.0
filelock==3.12.4
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gpustat==1.1.1
greenlet==3.0.1
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
Levenshtein==0.23.0
lit==17.0.3
llama_cpp_python==0.2.26
llvmlite==0.41.1
Mako==1.3.0
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2
ninja==1.11.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.3.8
optimum==1.13.2
optuna==3.4.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
peft==0.5.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.23.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.5.2
rdflib==7.0.0
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge==1.0.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.35.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.23
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.0
tensorboard-data-server==0.7.1
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.14.1
torch==2.0.1
torch_geometric==2.4.0
torchaudio==2.1.0
torchbiggraph @ git+https://github.com/joelbarmettlerUZH/PyTorch-BigGraph.git@7faa89d6ae034332b1836d2f6a0e51bdc931afe8
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.34.1
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
umap-learn==0.5.5
uri-template==1.3.0
urllib3==2.0.7
wandb==0.16.1
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.0
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.2
